requirements.txt
Django==4.2.16
djangorestframework==3.14.0
djoser==2.2.3
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-filter==24.3
Pillow==10.4.0
psycopg2-binary==2.9.9
gunicorn==23.0.0
python-dotenv==1.0.1
whitenoise==6.7.0
